+++ src/rics/translation/offline/_translation_map.py
@@ -45,17 +45,23 @@ class TranslationMap(Mapping, Generic[NameType, IdType, SourceType]):
         default_fmt: Optional[Format],
         default: Optional[DefaultTranslations[SourceType]],
     ) -> Dict[SourceType, FormatApplier]:
-        default = default or {}
+        dt: DefaultTranslations = default or DefaultTranslations()
 
-        if default_fmt is not None:
+        """if default_fmt is not None:
             # Ensure default translations are populated with something
             for source in source_placeholder_translations:
-                if source not in default:
-                    default[source] = {}
+                if source not in dt:
+                    dt[source] = {}"""
 
         return {
-            source: TranslationMap.FORMAT_APPLIER_TYPE(placeholder_translations, default.get(source, NO_DEFAULT))
-            for source, placeholder_translations in source_placeholder_translations.items()
+            source: TranslationMap.FORMAT_APPLIER_TYPE(
+                placeholder_translations=pht,
+                default=dt.get(source, NO_DEFAULT),
+                required_placeholders_for_default=pht.placeholders
+                if default_fmt is None
+                else default_fmt.required_placeholders,
+            )
+            for source, pht in source_placeholder_translations.items()
         }
 
     def apply(self, name: NameType, fmt: FormatType = None, default_fmt: FormatType = None) -> MagicDict[IdType]: